+++ lib/ecto/associations/assoc.ex
@@ -131,7 +131,7 @@ defmodule Ecto.Associations.Assoc do
   defp compare({pos1, _}, {pos2, _}), do: pos1 < pos2
 
   defp set_loaded(struct, refl, loaded) do
-    unless is_record(refl, Ecto.Reflections.HasMany) do
+    unless refl.__struct__ == Ecto.Reflections.HasMany do
       loaded = List.first(loaded)
     end
     Ecto.Associations.load(struct, refl.field, loaded)